requirements.txt
asgiref==3.6.0
asttokens==2.2.1
async-generator==1.10
attrs==23.1.0
backcall==0.2.0
beautifulsoup4==4.12.2
branca==0.6.0
certifi==2023.5.7
cffi==1.15.1
charset-normalizer==3.1.0
colorama==0.4.6
comm==0.1.3
contourpy==1.1.0
cycler==0.11.0
debugpy==1.6.7
decorator==5.1.1
Django==4.1.7
djangorestframework==3.14.0
emoji==2.6.0
exceptiongroup==1.1.1
executing==1.2.0
folium==0.14.0
fonttools==4.40.0
googlemaps==4.10.0
h11==0.14.0
idna==3.4
importlib-metadata==6.7.0
importlib-resources==5.12.0
install==1.3.5
ipykernel==6.23.3
ipython==8.14.0
jedi==0.18.2
Jinja2==3.1.2
joblib==1.3.0
JPype1==1.4.1
jupyter_client==8.3.0
jupyter_core==5.3.1
kiwisolver==1.4.4
konlpy==0.6.0
lxml==4.9.2
MarkupSafe==2.1.3
matplotlib==3.7.1
matplotlib-inline==0.1.6
nest-asyncio==1.5.6
numpy==1.25.0
outcome==1.2.0
packaging==23.1
pandas==2.0.2
parso==0.8.3
pickleshare==0.7.5
Pillow==9.5.0
platformdirs==3.8.0
prompt-toolkit==3.0.38
psutil==5.9.5
pure-eval==0.2.2
pycparser==2.21
Pygments==2.15.1
pyparsing==3.1.0
PySocks==1.7.1
python-dateutil==2.8.2
python-dotenv==1.0.0
pytz==2023.3
pywin32==306
pyzmq==25.1.0
requests==2.31.0
scikit-learn==1.2.2
scipy==1.11.1
seaborn==0.12.2
selenium==4.9.1
six==1.16.0
sniffio==1.3.0
sortedcontainers==2.4.0
soupsieve==2.4.1
sqlparse==0.4.3
stack-data==0.6.2
threadpoolctl==3.1.0
tornado==6.3.2
tqdm==4.65.0
traitlets==5.9.0
trio==0.22.0
trio-websocket==0.10.3
typing_extensions==4.7.0
tzdata==2023.3
urllib3==2.0.3
wcwidth==0.2.6
webdriver-manager==3.8.6
wsproto==1.2.0
zipp==3.15.0
